update quotes section compo
